+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Add `meraki_wireless_ssid_identity_psks` resource
 - Add `meraki_wireless_rf_profiles` resource
 - Add `meraki_network_webhook_payload_templates` resource
+- Add `meraki_switch_stack_routing_static_routes` resource
 
 ## 1.3.0